npm modules, vite - development server, electron forge
Vite is a local development server which is used for hot re-loading. Package.json contains all the dependencies and information about the repo like author's information etc. Index.html is the base file of react. requirements.txt contains the python modules that need to be imported for the use of the application. package-lock.json stores the exact versions required for the application. forge.config.js is the electron configuration file.